+++ lambdas/db-migration/main.tf
@@ -71,7 +71,7 @@ resource "aws_lambda_function" "db_migration" {
   role             = aws_iam_role.db_migration.arn
   runtime          = "nodejs12.x"
   timeout          = 120
-  memory_size      = 128
+  memory_size      = 256
 
   environment {
     variables = {